refactor!: replace debuglogger with apex
BREAKING CHANGE: this commit replaces the previous overly verbose logging solution with apex. All consumers will need to be able to use apex in order to control the log level. Overall it should however improve the usage of this package.

Mitigation: Remove debugLogger from the init of this package. If you want to control the log level import the apex log package into your project and set it there.